refactor(objectstore): separate sigv4 presign helpers
